Update BedrockifyClientSettings.java
There is a graphic glitch within the group screen of the simple voice mod from henkelmax. The glitch happens, if the setting for the "show panoramascreen" is activ in BedrockIfy. I figured out, if the entry "de.maxhenkel.voicechat" is added to the disabled list for the panorama setting, the issue is gone and the panorama setting can be keep enabled.

https://github.com/henkelmax/simple-voice-chat/issues/565